Extract toast view state
